// whoxa_frontend/src/types/ResType.ts
export interface OtpRes {
  message: string;
  success: boolean;
  token: string;
  resData: ResData;
}

export interface ResData {
  profile_image: string;
  user_id: number;
  phone_number: string;
  country: string;
  first_name: string;
  last_name: string;
  device_token: string;
  one_signal_player_id: string;
  user_name: string;
  bio: string;
  dob: string;
  country_code: string;
  password: string;
  last_seen: number;
  otp: number;
  gender: string;
  createdAt: Date;
  updatedAt: Date;
}